requirements.txt
allennlp==2.10.1
allennlp_models==2.10.1
fasttext==0.9.2
flair==0.11.1
gensim==4.1.2
jsonnet_binary==0.17.0
numpy==1.24.2
ray==2.0.0
regex==2021.8.3
tokenizers==0.11.6
torch==1.11.0
tqdm==4.62.3
transformers==4.17.0
